// ueu/libs/packages/sass-functions/src/lib/interfaces/token.model.ts
export enum TokenType {
    SPACE = 'space',
    COLOR = 'color',
    BRAND = 'brand',
    FONT_SIZE = 'fontSize',
    FONT_WEIGHT = 'fontWeight',
    LINE_HEIGHT = 'lineHeight',
    Z = 'z',
    TYPE = 'type',
    DEFAULT = '',
    LAYOUT = 'layout',
    OPACITY = 'opacity',
    TRANSITION = 'transition',
    BORDER_RADIUS = 'borderRadius',
    BORDER_WIDTH = 'borderWidth',
    BOX_SHADOW = 'boxShadow',
    SIZE = 'size',
}